requirements.txt
# Core dependencies
httpx>=0.27.0
fastapi>=0.111.0
PyJWT>=2.6.0
python-dotenv

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.20.0

# Build dependencies
setuptools>=68
wheel
